cuda: warning squashes with clang

+++ src/backend/cuda/pup/yaksuri_cudai_pup.c
@@ -11,35 +11,37 @@
 #include <stdlib.h>
 
 #define THREAD_BLOCK_SIZE  (256)
-#define MAX_GRIDSZ_X       ((1ULL << 31) - 1)
-#define MAX_GRIDSZ_Y       (65535)
-#define MAX_GRIDSZ_Z       (65535)
+#define MAX_GRIDSZ_X       ((1U << 31) - 1)
+#define MAX_GRIDSZ_Y       (65535U)
+#define MAX_GRIDSZ_Z       (65535U)
 
 #define MAX_IOV_LENGTH (16384)
 
-static int get_thread_block_dims(uint64_t count, yaksi_type_s * type, int *n_threads,
-                                 int *n_blocks_x, int *n_blocks_y, int *n_blocks_z)
+static int get_thread_block_dims(uintptr_t count, yaksi_type_s * type, unsigned int *n_threads,
+                                 unsigned int *n_blocks_x, unsigned int *n_blocks_y,
+                                 unsigned int *n_blocks_z)
 {
     int rc = YAKSA_SUCCESS;
     yaksuri_cudai_type_s *cuda_type = (yaksuri_cudai_type_s *) type->backend.cuda.priv;
 
     *n_threads = THREAD_BLOCK_SIZE;
-    uint64_t n_blocks = count * cuda_type->num_elements / THREAD_BLOCK_SIZE;
+    uintptr_t n_blocks = count * cuda_type->num_elements / THREAD_BLOCK_SIZE;
     n_blocks += ! !(count * cuda_type->num_elements % THREAD_BLOCK_SIZE);
 
     if (n_blocks <= MAX_GRIDSZ_X) {
-        *n_blocks_x = (int) n_blocks;
+        *n_blocks_x = (unsigned int) n_blocks;
         *n_blocks_y = 1;
         *n_blocks_z = 1;
     } else if (n_blocks <= MAX_GRIDSZ_X * MAX_GRIDSZ_Y) {
-        *n_blocks_x = YAKSU_CEIL(n_blocks, MAX_GRIDSZ_Y);
-        *n_blocks_y = YAKSU_CEIL(n_blocks, (*n_blocks_x));
+        *n_blocks_x = (unsigned int) (YAKSU_CEIL(n_blocks, MAX_GRIDSZ_Y));
+        *n_blocks_y = (unsigned int) (YAKSU_CEIL(n_blocks, (*n_blocks_x)));
         *n_blocks_z = 1;
     } else {
-        int n_blocks_xy = YAKSU_CEIL(n_blocks, MAX_GRIDSZ_Z);
-        *n_blocks_x = YAKSU_CEIL(n_blocks_xy, MAX_GRIDSZ_Y);
-        *n_blocks_y = YAKSU_CEIL(n_blocks_xy, (*n_blocks_x));
-        *n_blocks_z = YAKSU_CEIL(n_blocks, (uintptr_t) (*n_blocks_x) * (*n_blocks_y));
+        uintptr_t n_blocks_xy = YAKSU_CEIL(n_blocks, MAX_GRIDSZ_Z);
+        *n_blocks_x = (unsigned int) (YAKSU_CEIL(n_blocks_xy, MAX_GRIDSZ_Y));
+        *n_blocks_y = (unsigned int) (YAKSU_CEIL(n_blocks_xy, (*n_blocks_x)));
+        *n_blocks_z =
+            (unsigned int) (YAKSU_CEIL(n_blocks, (uintptr_t) (*n_blocks_x) * (*n_blocks_y)));
     }
 
   fn_exit:
@@ -151,8 +153,8 @@ int yaksuri_cudai_ipack(const void *inbuf, void *outbuf, uintptr_t count, yaksi_
         rc = yaksuri_cudai_md_alloc(type);
         YAKSU_ERR_CHECK(rc, fn_fail);
 
-        int n_threads;
-        int n_blocks_x, n_blocks_y, n_blocks_z;
+        unsigned int n_threads;
+        unsigned int n_blocks_x, n_blocks_y, n_blocks_z;
         rc = get_thread_block_dims(count, type, &n_threads, &n_blocks_x, &n_blocks_y, &n_blocks_z);
         YAKSU_ERR_CHECK(rc, fn_fail);
 
@@ -278,8 +280,8 @@ int yaksuri_cudai_iunpack(const void *inbuf, void *outbuf, uintptr_t count, yaks
         rc = yaksuri_cudai_md_alloc(type);
         YAKSU_ERR_CHECK(rc, fn_fail);
 
-        int n_threads;
-        int n_blocks_x, n_blocks_y, n_blocks_z;
+        unsigned int n_threads;
+        unsigned int n_blocks_x, n_blocks_y, n_blocks_z;
         rc = get_thread_block_dims(count, type, &n_threads, &n_blocks_x, &n_blocks_y, &n_blocks_z);
         YAKSU_ERR_CHECK(rc, fn_fail);
